+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.0
+* @BREAKING: Migrate to federated. removed static methods. You will need to create a class before using.
+* 
+* Min Flutter Version 3.10.6 due to shared swift code
+* MacOS support
 ## 0.1.1
 * Fix plugin crashes in background on Android.
 ## 0.1.0